pyproject.toml
[tool.poetry]
name = "my-deep-research"
version = "0.1.0"
description = "Deep Research using OpenAI agents SDK with series of different examples"
authors = ["Dmitry Y <your.email@example.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
openai = ">=1.12.0"
openai-agents = ">=0.0.14"
python-dotenv = ">=1.0.0"
beautifulsoup4 = ">=4.12.0"
aiohttp = ">=3.8.0"
tqdm = ">=4.65.0"
psycopg2-binary = ">=2.9.0"
pgvector = ">=0.2.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-mock = "^3.11.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
sequential-search = "src.sequential_search_agent.run:main" 